requirements.txt
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
python-dotenv==1.0.0
pytest>=7.0.0 # Added for unit testing
pytest-mock>=3.0.0 # Added for mocking dependencies in tests
